// src/types/news.ts
export interface Review {
    title: string;
    group?: string;
    date: string;
    type: string;
};

export interface Announcement {
    title: string;
    category?: string;
    date: string;
    type: string;
};

export interface NewsLetter {
    title: string;
    date: string;
    type: string;
}